feat: Initial flutter implementation
Adds the initial version of the Flutter moderated chat implementation.
It is partially styled and does not yet include image handling. The web
version doesn't work because the dart sdk doesn't support it, so any
testing should be done on the macOS version. I'll remove the web target
and add iOS and Android.

The important files are in the lib directory. The app consists of the
login_screen and chat_screen views. There are several services and
widgets that the views use to communicate with the back end.